move /register page to /sign-up
