+++ app/sign-up/__tests__/page.test.ts
