+++ app/sign-up/__tests__/success.test.ts
