+++ app/sign-up/_components/StepTracker.tsx
