+++ app/sign-up/_components/__tests__/AuthCardSkeleton.test.ts
