+++ app/sign-up/_components/__tests__/StepTracker.test.ts
